requirements.txt
fastapi
uvicorn
python-dotenv
openai
bs4
requests
sqlite3